+++ lib/src/config/routes/routes.dart
@@ -1,4 +1,4 @@
-import 'package:doft/src/presentation/register/registerscreen.dart';
+import 'package:doft/src/presentation/register/pages/register_screen.dart';
 import 'package:flutter/material.dart';
 
 import '../../presentation/home/home_screen.dart';
