+++ lib/src/core/utils/image_manager.dart
@@ -1,0 +1,4 @@
+class ImageManager {
+  static const String logo = 'assets/images/route.png';
+  static const String profile = 'assets/images/profile-circle.svg';
+}
